src/services/instance_manager/requirements.txt
grpcio-tools
protobuf
oauth2client
google-api-python-client